+++ src/vmm/src/lib.rs
@@ -348,7 +348,7 @@ impl Vmm {
 
         Vcpu::register_kick_signal_handler();
 
-        self.vcpus_handles.reserve(vcpu_count as usize);
+        self.vcpus_handles.reserve(vcpu_count);
 
         for mut vcpu in vcpus.drain(..) {
             vcpu.set_mmio_bus(self.mmio_device_manager.bus.clone());
