+++ src/vmm/src/persist.rs
@@ -303,8 +303,7 @@ fn snapshot_memory_to_file(
 
     // Set the length of the file to the full size of the memory area.
     let mem_size_mib = mem_size_mib(vmm.guest_memory());
-    // Set the length of the file to the full size of the memory area.
-    file.set_len((mem_size_mib * 1024 * 1024) as u64)
+    file.set_len(mem_size_mib * 1024 * 1024)
         .map_err(|err| MemoryBackingFile("set_length", err))?;
 
     match snapshot_type {
